Moved nested-layout demo to a seperate directory.
