+++ samples/dashboard/demos/dashboard-layout-nested/demo.css
@@ -1,0 +1,12 @@
+@import "https://code.highcharts.com/css/dashboard/gui.css";
+
+.col,
+[class*="col-"] {
+    padding-top: 1rem;
+    padding-bottom: 1rem;
+    border: 1px solid rgb(26 54 126 / 12.5%);
+}
+
+#container-bootstrap {
+    display: none;
+}
