+++ samples/dashboard/demos/dashboard-layout-nested/demo.details
@@ -1,0 +1,6 @@
+---
+ name: Highcharts Demo
+ authors:
+   - Sebastian Bochan, Wojciech Chmiel
+ js_wrap: b
+...
