+++ samples/dashboard/demos/dashboard-layout-nested/demo.html
@@ -1,0 +1,1 @@
+<div id="container" class="highcharts-dashboard"></div>
