+++ samples/dashboard/demos/dashboard-layout/demo.html
@@ -28,10 +28,6 @@
 <div id="container" class="highcharts-dashboard">
 </div>
 
-<h1>Dashboard layout in layout</h1>
-
-<div id="container-nested-layout" class="highcharts-dashboard"></div>
-
 <h1>Dashboard (User GUI - bootstrap)</h1>
