+++ samples/dashboard/demos/dashboard-layout/demo.mjs
@@ -596,301 +596,6 @@ const dashboardBootstrap = new Dashboard('container-bootstrap', {
 
 // console.log(dashboardBootstrap);
 
-console.log('========= Layout in layout =========');
-
-const dashboardLayout = new Dashboard('container-nested-layout', {
-    editMode: {
-        enabled: true,
-        contextMenu: {
-            icon: 'https://code.highcharts.com/gfx/dashboard-icons/menu.svg',
-            enabled: true,
-            items: ['editMode', {
-                id: 'export-dashboard',
-                text: 'Export dashboard',
-                events: {
-                    click: function () {
-                        dashboard.exportLocal();
-                    }
-                }
-            }, {
-                id: 'delete-dashboard',
-                text: 'Delete current dashboard',
-                events: {
-                    click: function () {
-                        dashboard.destroy();
-                    }
-                }
-            }, {
-                id: 'import-dashboard',
-                text: 'Import saved dashboard',
-                events: {
-                    click: function () {
-                        dashboard = Dashboard.importLocal();
-                    }
-                }
-            }, {
-                id: 'export-layout',
-                text: 'Export 1 layout',
-                events: {
-                    click: function () {
-                        exportedLayoutId = dashboard.layouts[0].options.id;
-                        dashboard.layouts[0].exportLocal();
-                    }
-                }
-            }, {
-                id: 'delete-layout',
-                text: 'Delete 1 layout',
-                events: {
-                    click: function () {
-                        dashboard.layouts[0].destroy();
-                    }
-                }
-            }, {
-                id: 'import-layout',
-                text: 'Import saved layout',
-                events: {
-                    click: function () {
-                        const layout = dashboard.importLayoutLocal(
-                            exportedLayoutId
-                        );
-                        console.log('Imported layout: ', layout);
-                    }
-                }
-            }]
-        },
-        toolbars: {
-            cell: {
-                menu: {
-                    items: [{
-                        id: 'drag',
-                        icon: 'https://code.highcharts.com/gfx/dashboard-icons/drag.svg'
-                    }, {
-                        id: 'settings',
-                        icon: 'https://code.highcharts.com/gfx/dashboard-icons/settings.svg'
-                    },
-                    // {
-                    //     id: 'my-option-1',
-                    //     text: 't1',
-                    //     events: {
-                    //         click: function() {
-                    //             console.log('hello world!');
-                    //         }
-                    //     }
-                    // },
-                    {
-                        id: 'destroy',
-                        icon: 'https://code.highcharts.com/gfx/dashboard-icons/destroy.svg'
-                    }]
-                }
-            },
-            row: {
-                menu: {
-                    items: [{
-                        id: 'drag',
-                        icon: 'https://code.highcharts.com/gfx/dashboard-icons/drag.svg'
-                    }, {
-                        id: 'settings',
-                        icon: 'https://code.highcharts.com/gfx/dashboard-icons/settings.svg'
-                    }, {
-                        id: 'destroy',
-                        icon: 'https://code.highcharts.com/gfx/dashboard-icons/destroy.svg'
-                    }]
-                }
-            },
-            settings: {
-                closeIcon: 'https://code.highcharts.com/gfx/dashboard-icons/close.svg',
-                dragIcon: 'https://code.highcharts.com/gfx/dashboard-icons/drag.svg'
-            }
-        },
-        lang: {
-            editMode: 'My edit mode',
-            chartOptions: 'Chart options EN'
-        },
-        tools: {
-            addComponentBtn: {
-                icon: 'https://code.highcharts.com/gfx/dashboard-icons/add.svg'
-            }
-        },
-        confirmationPopup: {
-            close: {
-                icon: 'https://code.highcharts.com/gfx/dashboard-icons/close.svg'
-            }
-        },
-        resize: {
-            enabled: true,
-            styles: {
-                minWidth: 20,
-                minHeight: 50
-            },
-            type: 'xy',
-            snap: {
-                width: 20,
-                height: 20
-            }
-        }
-    },
-    gui: {
-        enabled: true,
-        // layoutOptions: {
-        //     resize: {
-        //         cells: {
-        //             enabled: true,
-        //             minSize: 70
-        //         },
-        //         rows: {
-        //             enabled: true,
-        //             minSize: 70
-        //         }
-        //     }
-        // },
-        layouts: [{
-            id: 'layout-in-1', // mandatory
-            rows: [{
-                cells: [{
-                    id: 'dashboard-col-nolayout-0'
-                }, {
-                    id: 'dashboard-col-layout-0',
-                    layout: {
-                        rows: [{
-                            cells: [{
-                                id: 'dashboard-col-layout-1',
-                                width: '1/2'
-                            }, {
-                                id: 'dashboard-col-layout-2',
-                                layout: {
-                                    rows: [{
-                                        cells: [{
-                                            id: 'dashboard-col-layout-2a'
-                                        }, {
-                                            id: 'dashboard-col-layout-2b'
-                                        }, {
-                                            id: 'dashboard-col-layout-2f'
-                                        }]
-                                    }, {
-                                        cells: [{
-                                            id: 'dashboard-col-layout-2c'
-                                        }, {
-                                            id: 'dashboard-col-layout-2d'
-                                        }]
-                                    }]
-                                }
-                            }, {
-                                id: 'dashboard-col-layout-4',
-                                width: '1/2'
-                            }, {
-                                id: 'dashboard-col-layout-5',
-                                width: '1/3'
-                            }]
-                        }, {
-                            cells: [{
-                                id: 'dashboard-col-layout-3'
-                            }]
-                        }]
-                    }
-                }]
-            }]
-        }]
-    },
-    components: [{
-        cell: 'dashboard-col-nolayout-0',
-        type: 'Highcharts',
-        chartOptions: {
-            chart: {
-                animation: false
-            },
-            series: [{
-                name: 'Series from options',
-                data: [1, 2, 3, 4]
-            }]
-        }
-    }, {
-        cell: 'dashboard-col-layout-1',
-        ...chartDemo
-        // type: 'html',
-        // elements: [{
-        //     tagName: 'img',
-        //     attributes: {
-        //         src: 'https://i.ytimg.com/vi/qlO4M6MfDFY/hqdefault.jpg',
-        //         title: 'I heard you like components'
-        //     }
-        // }, {
-        //     textContent: 'Loreum ipsum'
-        // }]
-    }, {
-        cell: 'dashboard-col-layout-2a',
-        ...chartDemo
-        // type: 'html',
-        // elements: [{
-        //     tagName: 'img',
-        //     attributes: {
-        //         src: 'https://i.ytimg.com/vi/qlO4M6MfDFY/hqdefault.jpg',
-        //         title: 'I heard you like components'
-        //     }
-        // }, {
-        //     textContent: 'Loreum ipsum'
-        // }]
-    }, {
-        cell: 'dashboard-col-layout-2b',
-        ...chartDemo
-        // type: 'html',
-        // elements: [{
-        //     tagName: 'img',
-        //     attributes: {
-        //         src: 'https://i.ytimg.com/vi/qlO4M6MfDFY/hqdefault.jpg',
-        //         title: 'I heard you like components'
-        //     }
-        // }, {
-        //     textContent: 'Loreum ipsum'
-        // }]
-    }, {
-        cell: 'dashboard-col-layout-2c', ...chartDemo
-    }, {
-        cell: 'dashboard-col-layout-2d', ...chartDemo
-    }, {
-        cell: 'dashboard-col-layout-3',
-        ...chartDemo
-        // type: 'html',
-        // elements: [{
-        //     tagName: 'img',
-        //     attributes: {
-        //         src: 'https://i.ytimg.com/vi/qlO4M6MfDFY/hqdefault.jpg',
-        //         title: 'I heard you like components'
-        //     }
-        // }, {
-        //     textContent: 'Loreum ipsum'
-        // }]
-    }, {
-        cell: 'dashboard-col-layout-4',
-        ...chartDemo
-        // type: 'html',
-        // elements: [{
-        //     tagName: 'img',
-        //     attributes: {
-        //         src: 'https://i.ytimg.com/vi/qlO4M6MfDFY/hqdefault.jpg',
-        //         title: 'I heard you like components'
-        //     }
-        // }, {
-        //     textContent: 'Loreum ipsum'
-        // }]
-    }, {
-        cell: 'dashboard-col-layout-5',
-        ...chartDemo
-        // type: 'html',
-        // elements: [{
-        //     tagName: 'img',
-        //     attributes: {
-        //         src: 'https://i.ytimg.com/vi/qlO4M6MfDFY/hqdefault.jpg',
-        //         title: 'I heard you like components'
-        //     }
-        // }, {
-        //     textContent: 'Loreum ipsum'
-        // }]
-    }, {
-        cell: 'dashboard-col-layout-2f',
-        ...chartDemo
-    }]
-});
-
 // console.log(dashboardLayout);
 
 /*
